added delete and update options to flashcards and decks

+++ src/main/java/com/cpalmer/projects/flashcards/repository/DeckRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DeckRepository extends JpaRepository<Deck, Long> {
+public interface DeckRepository extends JpaRepository<Deck, Integer> {
     Deck findByDeckId(int deckId);
 }
